♻️ refactor: move _favourite.py to utils/favourite.py
Avoids confusion with extractor resource types. Update imports in
favourites.py and user_all_favourites.py accordingly.
