+++ src/yutto/extractor/favourites.py
@@ -8,11 +8,11 @@ from yutto.api.space import get_favourite_info, get_favourite_items, get_user_na
 from yutto.api.ugc_video import get_ugc_video_list
 from yutto.exceptions import NoAccessPermissionError, NotFoundError
 from yutto.extractor._abc import BatchExtractor
-from yutto.extractor._favourite import normalize_favourite_video_item
 from yutto.extractor.common import extract_ugc_video_data
 from yutto.types import FId, MId
 from yutto.utils.asynclib import CoroutineWrapper
 from yutto.utils.console.logger import Badge, Logger
+from yutto.utils.favourite import normalize_favourite_video_item
 from yutto.utils.fetcher import Fetcher
 from yutto.utils.filter import Filter
