+++ src/yutto/utils/favourite.py
